backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
mysqlclient>=2.2
gunicorn>=21.2
python-dotenv~=1.2.1